+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v5.5.0
 
 **FieldConvert**
 - Fix typo in user-guide and fix but for parallel-in-time FieldConvert (!1645)
+- Fixed FieldConvert -m addfld (!1500)
 
 v5.4.0
 ------